refactor: 522 - rename social-media to social-media-item
