+++ src/shared/ui/social-media-item/index.ts
@@ -1,0 +1,1 @@
+export { SocialMediaItem, type SocialMediaProps } from './social-media-item';
