+++ src/shared/ui/social-media-item/social-media-item.module.scss
