--- src/shared/ui/social-media/index.ts
@@ -1,1 +1,0 @@
-export { SocialMedia, type SocialMediaProps } from './social-media';
